requirements.txt
numpy
matplotlib
seaborn
tensorflow
keras
streamlit
streamlit-drawable-canvas
pillow